jahs_fashion/requirements.txt
absl-py==1.2.0
astunparse==1.6.3
attrs==22.1.0
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0
ConfigSpace==0.4.21
Cython==0.29.32
dask==2022.8.1
dehb==0.0.3
dill==0.3.5.1
distributed==2022.8.1
emcee==3.1.2
flatbuffers==1.12
fsspec==2022.7.1
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.47.0
h5py==3.7.0
HeapDict==1.0.1
icgen @ git+https://github.com/automl/ICGen.git@c04bf220f24fd475906225ca2c9176949cd6f129
idna==3.3
importlib-metadata==4.12.0
iopath==0.1.10
jahs-bench-201==0.1.0
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
locket==1.0.0
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
more-itertools==8.14.0
msgpack==1.0.4
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
partd==1.3.0
Pillow==7.2.0
portalocker==2.5.1
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynisher==1.0.0
pyparsing==3.0.9
pyrfr==0.8.3
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.9.0
six==1.16.0
smac==1.2
sortedcontainers==2.4.0
tabulate==0.8.10
tblib==1.7.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-datasets==2.1.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.10.0
termcolor==1.1.0
threadpoolctl==3.1.0
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchsummary==1.5.1
torchvision==0.13.1
tornado==6.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
xgboost==1.5.2
yacs==0.1.8
zict==2.2.0
zipp==3.8.1
